refactor: make one class encompassing all fields of gibsons form
Move all fields to a single class GibsonsForm, and make CollectionState only carry one instance of this class.

BREAKING CHANGE:
