--- lib/collection/models/interview_start_time.dart
@@ -1,17 +1,0 @@
-import 'package:formz/formz.dart';
-
-enum InterviewStartTimeValidationError { invalid }
-
-class InterviewStartTime
-    extends FormzInput<String, InterviewStartTimeValidationError> {
-  const InterviewStartTime.pure() : super.pure('');
-  const InterviewStartTime.dirty([String value = '']) : super.dirty(value);
-
-  @override
-  InterviewStartTimeValidationError? validator(String? value) {
-    // TODO: Add validation, currently only checks if not empty
-    return value?.isNotEmpty == true
-        ? null
-        : InterviewStartTimeValidationError.invalid;
-  }
-}
